Create structure for browser tests
