+++ package.json
@@ -10,11 +10,10 @@
     "build": "browserify -s 'storj' -d ./lib/browser.js > storj.es6.debug.js && babel -o storj.es5.js -s true ./storj.es6.debug.js && cat storj.es6.debug.js | exorcist storj.es6.js.map > storj.es6.js && rm ./storj.es6.debug.js",
     "posttest": "npm run linter",
     "test": "npm run test-node && npm run test-browser",
-    "test-node": "nyc --reporter=text --reporter=html tape test/**/*.js",
+    "test-node": "nyc --reporter=text --reporter=html tape test/all/*.js",
     "test-browser": "testling",
     "linter": "eslint .",
-    "make-docs": "rm -rf ./jsdoc && ./node_modules/.bin/jsdoc lib -r -R README.md -c .jsdoc.json --verbose -d ./jsdoc",
-    "node-test": "node test/**/*.js"
+    "make-docs": "rm -rf ./jsdoc && ./node_modules/.bin/jsdoc lib -r -R README.md -c .jsdoc.json --verbose -d ./jsdoc"
   },
   "browserify": {
     "transform": [
@@ -24,7 +23,7 @@
   "testling": {
     "files": [
       "lib/browser.js",
-      "test/**/*.js"
+      "test/{all,browser}/*.js"
     ]
   },
   "babel": {
